+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add `Fiber#mailbox`
+
 # 2025-05-05 Version 0.9
 
 - Add `#statx`